Function for getting list of already scraped bulletins created. Check for data directory added. Filename and path constants created.

+++ vvo.py
@@ -9,18 +9,26 @@ import os
 
 
 ROOT_URL = "http://www.uvo.gov.sk/evestnik/-/vestnik"
+BULLETIN_CATALOG = "bulletins.json"
+DATA_DIR = "data"
+
+
+def get_scraped_bulletins():
+    """Get list of already scraped bulletins from bulletin catalog file"""
+
+    original_bulletins = []
+    if os.path.exists(BULLETIN_CATALOG):
+        with open(BULLETIN_CATALOG, "r") as bulletin_file:
+            original_bulletins = json.load(bulletin_file)
+
+    return original_bulletins  
 
 
 def get_new_bulletins():
     """Scrape new bulletins"""
-    # get list of all already scraped bulletins from bulletin file
-    # if file does not exists, no bulletins were previously scraped
-    original_bulletins = []
-    try:
-        with open("bulletins.json", "r") as bulletin_file:
-            original_bulletins = json.load(bulletin_file)
-    except:
-        pass
+
+    # get list of all already scraped bulletins
+    scraped_bulletins = get_scraped_bulletins()
     
     # create connection
     responce = requests.get(ROOT_URL + "/all")
@@ -38,25 +46,30 @@ def get_new_bulletins():
                 "url": bulletin["href"]
             }
             # bulletin has not been scraped before, add it to unscraped list
-            if bulletin["issue"] not in original_bulletins:
+            if bulletin["issue"] not in scraped_bulletins:
                 bulletins.append(bulletin)
     return bulletins
 
 
 def download_xml(url):
     """Download xml file from announcement url. The url contains a numeric 
-       name of the xml file, no need to search the url content."""
+    name of the xml file, no need to search the url content."""
 
     # get file name from
     filename = "%s.xml" % (re.search("/([0-9]+);", url).group(1))
     url = ROOT_URL + "/save/" + filename
-    destination = os.path.join("data", filename)
+
+    # check if data directory exists, if not raise exception
+    if not os.path.exists(DATA_DIR):
+        raise Exception("Data directory does not exist")
+
+    destination = os.path.join(DATA_DIR, filename)
     urllib.urlretrieve(url, destination)
 
 
 def scrape_bulletin(bulletin):
-    """Scrape announcements from bulletin url as xml file.
-       Store bulletin issue into bulletin catalog file."""
+    """Scrape announcements from bulletin url as xml file. 
+    Store bulletin issue into bulletin catalog file."""
 
     # create connection
     response = requests.get(bulletin["url"])
@@ -68,17 +81,12 @@ def scrape_bulletin(bulletin):
     for announcement in announcements:
         download_xml(announcement["href"])
         
-    # get list of already scraped bulletins
-    bulletins = []
-    try:
-        with open("bulletins.json", "r") as bulletin_file:
-            bulletins = json.load(bulletin_file)
-    except:
-        pass
+    # get list of already scraped bulletins and add newly scraped one
+    bulletins = get_scraped_bulletins()
+    bulletins.append(bulletin["issue"])
     
     # store all scraped bulletins
-    with open("bulletins.json", "w") as bulletin_file:
-        bulletins.append(bulletin["issue"])
+    with open(BULLETIN_CATALOG, "w") as bulletin_file:
         json.dump(bulletins, bulletin_file, indent=4)
